Move get_interface_state() to module_utils
