+++ tox.ini
@@ -13,7 +13,7 @@ deps =
 [base]
 passenv = *
 setenv =
-    PYTHONPATH = {toxinidir}/library
+    PYTHONPATH = {toxinidir}/library:{toxinidir}/module_utils
     LC_ALL = C
 changedir = {toxinidir}/tests
 covtarget =
